4th Commit
Product search is ready.
I fixed the redirect when clicking on the cart product title.
The cart now empties when the voucher is released.
I added a component to window for payment when it failures and cancellations.
I also added a button to the receipt to "Download receipt", "Print receipt" and "Return to start" in voutcher.